+++ src/main/java/com/gasstation/managementsystem/service/impl/PumpCodeServiceImpl.java
@@ -1,18 +1,10 @@
 package com.gasstation.managementsystem.service.impl;
 
-import com.gasstation.managementsystem.entity.Card;
-import com.gasstation.managementsystem.entity.Pump;
 import com.gasstation.managementsystem.entity.PumpCode;
-import com.gasstation.managementsystem.entity.Shift;
-import com.gasstation.managementsystem.exception.custom.CustomNotFoundException;
-import com.gasstation.managementsystem.model.dto.pumpCode.PumpCodeDTO;
-import com.gasstation.managementsystem.model.dto.pumpCode.PumpCodeDTOCreate;
-import com.gasstation.managementsystem.model.dto.pumpCode.PumpCodeDTOUpdate;
-import com.gasstation.managementsystem.model.mapper.PumpCodeMapper;
+import com.gasstation.managementsystem.model.dto.PumpCodeDTO;
 import com.gasstation.managementsystem.repository.PumpCodeRepository;
 import com.gasstation.managementsystem.service.PumpCodeService;
-import com.gasstation.managementsystem.utils.OptionalValidate;
-import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -22,78 +14,42 @@ import java.util.HashMap;
 import java.util.List;
 
 @Service
-@RequiredArgsConstructor
 public class PumpCodeServiceImpl implements PumpCodeService {
-    private final PumpCodeRepository pumpCodeRepository;
-    private final OptionalValidate optionalValidate;
-
-    private HashMap<String, Object> listPumpCodeToMap(List<PumpCode> pumpCodes) {
-        List<PumpCodeDTO> pumpCodeDTOS = new ArrayList<>();
-        for (PumpCode pumpCode : pumpCodes) {
-            pumpCodeDTOS.add(PumpCodeMapper.toPumpCodeDTO(pumpCode));
-        }
-        HashMap<String, Object> map = new HashMap<>();
-        map.put("data", pumpCodeDTOS);
-        return map;
-    }
+    @Autowired
+    PumpCodeRepository pumpCodeRepository;
 
     @Override
     public HashMap<String, Object> findAll(Pageable pageable) {
         Page<PumpCode> pumpCodes = pumpCodeRepository.findAll(pageable);
-        HashMap<String, Object> map = listPumpCodeToMap(pumpCodes.getContent());
+        List<PumpCodeDTO> pumpCodeDTOS = new ArrayList<>();
+        for (PumpCode pumpCode : pumpCodes) {
+            pumpCodeDTOS.add(new PumpCodeDTO(pumpCode));
+        }
+        HashMap<String, Object> map = new HashMap<>();
+        map.put("data", pumpCodeDTOS);
         map.put("totalElement", pumpCodes.getTotalElements());
         map.put("totalPage", pumpCodes.getTotalPages());
         return map;
     }
 
     @Override
-    public HashMap<String, Object> findAll() {
-        return listPumpCodeToMap(pumpCodeRepository.findAll());
+    public PumpCodeDTO findById(int id) {
+        return new PumpCodeDTO(pumpCodeRepository.findById(id).get());
     }
 
     @Override
-    public PumpCodeDTO findById(int id) throws CustomNotFoundException {
-        return PumpCodeMapper.toPumpCodeDTO(optionalValidate.getPumpCodeById(id));
+    public PumpCodeDTO save(PumpCode pumpCode) {
+        pumpCodeRepository.save(pumpCode);
+        return new PumpCodeDTO(pumpCode);
     }
 
     @Override
-    public PumpCodeDTO create(PumpCodeDTOCreate pumpCodeDTOCreate) throws CustomNotFoundException {
-        PumpCode pumpCode = PumpCodeMapper.toPumpCode(pumpCodeDTOCreate);
-        Pump pump = optionalValidate.getPumpById(pumpCodeDTOCreate.getPumpId());
-        Shift shift = optionalValidate.getShiftById(pumpCodeDTOCreate.getShiftId());
-        Card card = optionalValidate.getCardById(pumpCodeDTOCreate.getCardId());
-        pumpCode.setPump(pump);
-        pumpCode.setShift(shift);
-        pumpCode.setCard(card);
-        pumpCode = pumpCodeRepository.save(pumpCode);
-        return PumpCodeMapper.toPumpCodeDTO(pumpCode);
-    }
-
-    @Override
-    public PumpCodeDTO update(int id, PumpCodeDTOUpdate pumpCodeDTOUpdate) throws CustomNotFoundException {
-        PumpCode pumpCode = optionalValidate.getPumpCodeById(id);
-        PumpCodeMapper.copyNonNullToFuel(pumpCode, pumpCodeDTOUpdate);
-        Integer pumpId = pumpCodeDTOUpdate.getPumpId();
-        Integer shiftId = pumpCodeDTOUpdate.getShiftId();
-        Integer cardId = pumpCodeDTOUpdate.getCardId();
-        if (pumpId != null) {
-            pumpCode.setPump(optionalValidate.getPumpById(pumpId));
+    public PumpCodeDTO delete(int id) {
+        PumpCode pumpCode = pumpCodeRepository.findById(id).get();
+        if (pumpCode != null) {
+            pumpCodeRepository.delete(pumpCode);
+            return new PumpCodeDTO(pumpCode);
         }
-        if (shiftId != null) {
-            pumpCode.setShift(optionalValidate.getShiftById(shiftId));
-        }
-        if (cardId != null) {
-            pumpCode.setCard(optionalValidate.getCardById(cardId));
-        }
-        pumpCode = pumpCodeRepository.save(pumpCode);
-        return PumpCodeMapper.toPumpCodeDTO(pumpCode);
-    }
-
-
-    @Override
-    public PumpCodeDTO delete(int id) throws CustomNotFoundException {
-        PumpCode pumpCode = optionalValidate.getPumpCodeById(id);
-        pumpCodeRepository.delete(pumpCode);
-        return PumpCodeMapper.toPumpCodeDTO(pumpCode);
+        return null;
     }
 }
